requirements.txt
groq
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
PyPDF2
pytest
